pyproject.toml
[tool.poetry]
name = "habit-trecker-bot-test"
version = "0.1.0"
description = ""
authors = ["Sprite_Spirit <74331153+SpriteSpirit@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytelegrambotapi = "^4.22.0"
django = "^5.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
permission = "^0.4.1"
djangorestframework-simplejwt = "^5.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
